# docs/verity/openapi.json
{
  "components": {
    "schemas": {
...
      "AegisEditApplyRequest": {
        "additionalProperties": false,
        "properties": {
          "apply": {
            "title": "Apply",
            "type": "boolean"
          },
          "edit_id": {
            "title": "Edit Id",
            "type": "string"
          }
        },
        "required": [
          "edit_id",
          "apply"
        ],
        "title": "AegisEditApplyRequest",
        "type": "object"
      },
      "AegisEditRequest": {
        "additionalProperties": false,
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "edit_types": {
            "items": {
              "$ref": "#/components/schemas/EditType"
            },
            "minItems": 1,
            "title": "Edit Types",
            "type": "array"
          },
          "locked_spans": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "maxItems": 1000,
            "title": "Locked Spans",
            "type": "array"
          },
          "operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "preserve_voice": {
            "default": true,
            "title": "Preserve Voice",
            "type": "boolean"
          },
          "span_end": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span End"
          },
          "span_start": {
            "default": 0,
            "minimum": 0.0,
            "title": "Span Start",
            "type": "integer"
          },
          "text": {
            "maxLength": 250000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "document_id"
        ],
        "title": "AegisEditRequest",
        "type": "object"
      },
      "AegisEditResponse": {
        "properties": {
          "ai_generated": {
            "title": "Ai Generated",
            "type": "boolean"
          },
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "change_reason": {
            "title": "Change Reason",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dimension": {
            "title": "Dimension",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_version_id": {
            "title": "Document Version Id",
            "type": "string"
          },
          "edit_type": {
            "$ref": "#/components/schemas/EditType"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "original": {
            "title": "Original",
            "type": "string"
          },
          "preserve_voice": {
            "title": "Preserve Voice",
            "type": "boolean"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "span_end": {
            "title": "Span End",
            "type": "integer"
          },
          "span_start": {
            "title": "Span Start",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "edit_type",
          "dimension",
          "original",
          "revision",
          "change_reason",
          "span_start",
          "span_end",
          "applied",
          "ai_generated",
          "document_id",
          "document_version_id",
          "preserve_voice",
          "engine_version",
          "created_at"
        ],
        "title": "AegisEditResponse",
        "type": "object"
      },
      "AegisHistoryResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AegisEditResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "document_id",
          "document_version_id",
          "items"
        ],
        "title": "AegisHistoryResponse",
        "type": "object"
      },
      "AegisRefineRequest": {
        "additionalProperties": false,
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "edit_types": {
            "items": {
              "$ref": "#/components/schemas/EditType"
            },
            "minItems": 1,
            "title": "Edit Types",
            "type": "array"
          },
          "locked_spans": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "maxItems": 1000,
            "title": "Locked Spans",
            "type": "array"
          },
          "operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "preserve_voice": {
            "default": true,
            "title": "Preserve Voice",
            "type": "boolean"
          },
          "text": {
            "maxLength": 250000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "document_id"
        ],
        "title": "AegisRefineRequest",
        "type": "object"
      },
      "AegisRefineResponse": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/AegisChangeResponse"
            },
            "title": "Changes",
            "type": "array"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_version_id": {
            "title": "Document Version Id",
            "type": "string"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          },
          "revised_text": {
            "title": "Revised Text",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "document_version_id",
          "original_text",
          "revised_text",
          "changes",
          "disclaimer"
        ],
        "title": "AegisRefineResponse",
        "type": "object"
      },
...
      "CreatedKey": {
        "properties": {
          "key": {
            "$ref": "#/components/schemas/KeyMetadata"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "key",
          "secret"
        ],
        "title": "CreatedKey",
        "type": "object"
      },
...
      "EditType": {
        "enum": [
          "grammar",
          "clarity",
          "concision",
          "tone",
          "structure",
          "citation",
          "vocabulary",
          "coherence"
        ],
        "title": "EditType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeyMetadata": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "rotated_from_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotated From Id"
          },
          "scopes": {
            "items": {
              "enum": [
                "text:analyze",
                "text:refine",
                "text:verify",
                "documents:read",
                "documents:write",
                "usage:read"
              ],
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "user_id",
          "organization_id",
          "scopes",
          "created_at",
          "expires_at",
          "last_used_at",
          "revoked_at",
          "is_active",
          "rotated_from_id"
        ],
        "title": "KeyMetadata",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AzaeronAPIKey": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Privacy-first writing intelligence with reviewable evidence",
    "title": "AZAERON VERITY",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/aegiswrite/apply": {
      "post": {
        "description": "Record an authenticated accept/reject decision for a persisted edit.",
        "operationId": "apply_edit_api_v1_aegiswrite_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisEditApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisEditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Apply Edit",
        "tags": [
          "AZAERON WRITE"
        ]
      }
    },
    "/api/v1/aegiswrite/history/{document_id}": {
      "get": {
        "description": "Return the tenant-scoped editorial ledger without exposing other workspaces.",
        "operationId": "edit_history_api_v1_aegiswrite_history__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "document_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit History",
        "tags": [
          "AZAERON WRITE"
        ]
      }
    },
    "/api/v1/aegiswrite/refine": {
      "post": {
        "description": "Return a deterministic editorial pass plus a persisted change ledger.",
        "operationId": "refine_draft_api_v1_aegiswrite_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisRefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisRefineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Refine Draft",
        "tags": [
          "AZAERON WRITE"
        ]
      }
    },
    "/api/v1/aegiswrite/suggest": {
      "post": {
        "description": "Persist transparent, version-bound suggestions for a selected text range.",
        "operationId": "suggest_edits_api_v1_aegiswrite_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AegisEditResponse"
                  },
                  "title": "Response Suggest Edits Api V1 Aegiswrite Suggest Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Suggest Edits",
        "tags": [
          "AZAERON WRITE"
        ]
      }
    },
    "/api/v1/api-keys": {
      "get": {
        "operationId": "listing_api_v1_api_keys_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KeyMetadata"
                  },
                  "title": "Response Listing Api V1 Api Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listing",
        "tags": [
          "API keys"
        ]
      },
      "post": {
        "operationId": "create_api_v1_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedKey"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create",
        "tags": [
          "API keys"
        ]
      }
    },
    "/api/v1/api-keys/{key_id}": {
      "delete": {
        "operationId": "revoke_api_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke",
        "tags": [
          "API keys"
        ]
      }
    },
    "/api/v1/api-keys/{key_id}/rotate": {
      "post": {
        "operationId": "rotate_api_v1_api_keys__key_id__rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedKey"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rotate",
        "tags": [
          "API keys"
        ]
      }
    },
...
  }
}